Particle stuff

+++ common/threading.cpp
@@ -48,7 +48,7 @@ idSysThread::~idSysThread() {
 idSysThread::StartThread
 ========================
 */
-bool idSysThread::StartThread( const char *name_, core_t core, xthreadPriority priority, int stackSize ) {
+bool idSysThread::StartThread( const char *name_, core_t core, xthreadPriority priority, size_t stackSize ) {
 	if ( isRunning ) {
 		return false;
 	}
@@ -72,7 +72,7 @@ bool idSysThread::StartThread( const char *name_, core_t core, xthreadPriority p
 idSysThread::StartWorkerThread
 ========================
 */
-bool idSysThread::StartWorkerThread( const char *name_, core_t core, xthreadPriority priority, int stackSize ) {
+bool idSysThread::StartWorkerThread( const char *name_, core_t core, xthreadPriority priority, size_t stackSize ) {
 	if ( isRunning ) {
 		return false;
 	}
